用餐信息预设与更改，购物车模块，下单功能，用户画像管理初版

+++ src/backend/src/main/java/com/fire/firepalaceorderingsys/dto/AiRecommendLogDTO.java
@@ -61,9 +61,9 @@ public class AiRecommendLogDTO {
     private BigDecimal actualOrderAmount;
 
     /**
-     * 推荐接受率
+     * 是否接受推荐：0-不接受，1-接受
      */
-    private Float acceptRate;
+    private Integer acceptRate;
 
     /**
      * 创建时间
